+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2026-02-09
+
+### Added
+
+- Group files by personal/team sections in dashboard Files tab
+- Show git repo URLs in file section headers
+
 ## [1.7.1] - 2026-02-08
 
 ### Added